% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aa_ac}
\alias{aa_ac}
\title{Aminoacids and Anticodons}
\format{Named (aminoacid and anticodon) vector of numerical (tRNA gene count) set to 0}
\source{
\url{http://www.tef-research.org/}
}
\usage{
aa_ac
}
\description{
Named Vector with Aminoacid and Anticodon Tokens with empty (0) counts
}
\keyword{datasets}

