% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anticodoncount.R
\name{anticodoncount}
\alias{anticodoncount}
\title{Calculate Global Anticodon Usage}
\usage{
anticodoncount(entry = 3)
}
\arguments{
\item{entry}{Index number in SeqFastadna list myaa}

\item{accnt}{Vector of 65 anticodons used}
}
\value{
Codon Frequency Table
}
\description{
\code{anticodoncount} sums up the anticodon usage in a mRNA and a protein fasta file
}
\details{
This function should count the codon frequency in a genome based on a set of verified and matched transcript and protein fasta files.
This is a roxygen2 created helping file.
}
\examples{
anticodoncount()

}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}}
}
\keyword{CodonBias}

