% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areopts.R
\name{areopts}
\alias{areopts}
\title{Optimality Decision for Codons of Custom Genome}
\usage{
areopts(entry = 3, opt = TRUE)
}
\arguments{
\item{entry}{Index number in SeqFastadna list myaa}

\item{opt}{If boolean and not fraction of tRNA genes to optimal isocode is used}

\item{areopt}{Vector of anticodon optimality  0>=areopt<=1}
}
\value{
Boolean or numerical vector with Optimality decision or score
}
\description{
\code{areopts} Calculates an Optimality Score for Codon in a Protein-Transcript Pair based on tRNA-Gene Abundancy in a Custom Genome (eg \emph{Eragostis tef})
}
\details{
This function should count the codon frequency in a genome based on a set of verified and matched transcript and protein fasta files.
This is a roxygen2 created helping file.
}
\examples{
areopts(entry=1, opt=TRUE)

}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}}
}
\keyword{CodonBias}

