% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCDS.R
\name{checkCDS}
\alias{checkCDS}
\title{Checks Integrity of CDS Sequence}
\usage{
checkCDS(cds)
}
\arguments{
\item{cds}{Sequence to be checked     # may be implemented as list List of Coding Sequences as read.fasta input}
}
\value{
TRUE if all checks are passed with or without warnings, FALSE if an violation is found
}
\description{
\code{checkCDS} Checks that the character number in the cds is in codons (modulo 3 = zero)
}
\details{
Based on early codon usage mesures proposed by Ikemura 1981. Original implementation from Alexander Roth 2005-2007,
first of several indices for codon usage.
}
\examples{
checkCDS(c2s('ATGTGGTACTCCGACTACGGAGGATAA'))

}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}} \code{\link{ComputeFop}}
}
\keyword{CodonBias}

