% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonProb.R
\name{CodonProb}
\alias{CodonProb}
\title{Codon Probability table}
\usage{
CodonProb(pos = 65)
}
\arguments{
\item{pos}{Which Codon Probabilities should be printed out}
}
\value{
Numerical vector with relative adaptiveness for the 64 codons
}
\description{
\code{CodonProb} creates a table of codon probabilities
}
\details{
Generation of manually set tables from literature. Maybe enhanced to compute for the used database.
}
\examples{
CodonProb(1:64)

}
\author{
Roth, A.; Friberg, M.; Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{statanacoseq}} \code{\link{SetupRA}}
}
\keyword{CodonBias}

