% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonUsage.R
\name{CodonUsage}
\alias{CodonUsage}
\title{Table for Codon Usage and Optimality of Codons}
\usage{
CodonUsage(cdslist = mylist())
}
\arguments{
\item{cdslist}{List of coding sequences in reading frame (eg. of full genome), handles also DNA string sequences.}
}
\value{
data frame with codon usage count Table with statistics for optimality based on occurence
}
\description{
\code{CodonUsage} calculates ratio of optimal codons used to the sum of synonymous codons
}
\details{
Creates a count table to decide on optimality of codons based on the usage of the imput DNA sequences.
}
\examples{
CodonUsage('ATGTGGTACTCCGACTACGGAGGATAA')
CodonUsage(mylist(whatout=1))

}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{statanacoseq}} \code{\link{ComputeFop}}
}
\keyword{CodonBias}

