% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCAIVector.R
\name{ComputeCAIVector}
\alias{ComputeCAIVector}
\title{Codon Adaptation Index Vector}
\usage{
ComputeCAIVector(cds, RA)
}
\arguments{
\item{cds}{Coding sequence in reading frame}
}
\value{
Vector of all 65 codon values
}
\description{
\code{ComputeCAIVector} Computes Codon Adaptation Index
}
\details{
Should compute the same ComputeCAI or ComputeCAIVector as in Darwin and the cai of seqinr package.
}
\section{Original code in Darwin}{

\subsection{Compute CAI, the Codon Adaptation Index (Sharp and Li 1987)}{\preformatted{
ComputeCAIVector := proc(e:Entry)
 if not assigned(RA) then
   error('Error in ComputeCAI: RA not assigned, use e.g. SetupRA(yeast);') fi;
 dna := SearchTag('DNA', e);
 wa := CreateArray(1..20);
 na := CreateArray(1..20);
 for j to length(dna) by 3 do
   cint := CodonToCInt(dna[j..j+2]);
   a := CIntToInt(cint);
   if a <= 20 then
     wa[a] := wa[a] + ln(RA[cint]);
     na[a] := na[a]+1;
   fi;
 od;
 res := CreateArray(1..21);
 for i to 20 do
   res[i] := If(na[i]=0, 'NA', exp(1/na[i] * wa[i])) od;
 res[21] := exp(1/sum(na) * sum(wa));
 res
end: } }
}
\examples{
ComputeCAIVector('ATGTGGTACTCCGACTACGGAGGATAA', RA=SetupRA("yeast"))
ComputeCAIVector(c2s(mylist(whatout=1)[[1]]), RA=ComputeCarboneRA(DB=mylist(whatout=1)))

}
\author{
Roth, A.; Friberg, M.; Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link[seqinr]{cai}} \code{\link{statanacoseq}} \code{\link{checkCDS}}
}
\keyword{CodonBias}

