% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCarboneRA.R
\name{ComputeCarboneRA}
\alias{ComputeCarboneRA}
\title{Compute Carbone Relative Adaptiveness}
\usage{
ComputeCarboneRA(t = 0.01, initfrac = 1, iterfrac = 0.5,
  mode = "reverse")
}
\arguments{
\item{t}{nonnegative}

\item{initfrac}{nonnegative}

\item{mode}{"reverse" for standard anything else for non-standard}

\item{interfrac}{nonnegative}
}
\value{
Named (codons) numerical vector with relative synonymous codon usage for the 64 codons
}
\description{
\code{ComputeCarboneRA} calculates RA
}
\details{
Should compute the same RA in Darwin
}
\examples{
ComputeCarboneRA()

}
\author{
Roth, A.; Friberg, M.; Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{RelativeAdaptiveness}} \code{\link{statanacoseq}} \code{\link{SetupRA}}
}
\keyword{CodonBias}

