% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCBI.R
\name{ComputeCBI}
\alias{ComputeCBI}
\title{Codon Bias Index}
\usage{
ComputeCBI(cds)
}
\arguments{
\item{cds}{Coding Sequence}
}
\value{
VOID
}
\description{
\code{ComputeCBI} Computes the Codon Bias Index for a Gene after Bennetzen and Hall 1982
}
\details{
Should be implemented because missing in Darwin
}
\section{Original code in Darwin}{

\subsection{Codon Bias Index (CBI) (Bennetzen and Hall 1982).}{\preformatted{
ComputeCBI :=  proc(d:string)
# to be implemented
end: } }
}
\examples{
ComputeCBI('ATGTGGTACTCCGACTACGGAGGATAA')
ComputeCBI(c2s(mylist(whatout=1)[[1]]))

}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}}
}
\keyword{CodonBias}

