% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeFop.R
\name{ComputeFop}
\alias{ComputeFop}
\title{Frequency of Optimal Codons}
\usage{
ComputeFop(cds, db, ref = c("tRNAgene", "mostcommon"), codonusagec = NULL)
}
\arguments{
\item{cds}{Coding sequence in reading frame}

\item{db}{CDS list to generate codon usage count table if NULL}

\item{ref}{'tRNAgene' takes tRNA gene copy number of isotypes out of genomic data given in codonusagec and 'mostcommon' takes
optimal codons as those that are the most common in a database or calculates it from a set of reference genes given as db}

\item{codonusagec}{Codon usage count table}
}
\value{
Numerical value for ratio of optimal codons (total number of optimal codons / total number of codons included in the analysis)
}
\description{
\code{ComputeFop} calculates ratio of optimal codons used to the sum of synonymous codons
}
\details{
Based on early codon usage mesures proposed by Ikemura 1981. Original implementation from Alexander Roth 2005-2007,
first of several indices for codon usage.
}
\examples{
ComputeFop('ATGTGGTACTCCGACTACGGAGGATAA', ref='tRNAgene', codonusagec=Tef)

}
\author{
Roth, A.; Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}} \url{http://gtrnadb.ucsc.edu/GtRNAdb2/} \code{\link{readstats}}
}
\keyword{CodonBias}

