% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeGC12syn.R
\name{ComputeGC12syn}
\alias{ComputeGC12syn}
\title{GC content 1st and 2nd position of synonymous codons}
\usage{
ComputeGC12syn(tD)
}
\arguments{
\item{tD}{Nucleotide character string}
}
\value{
vector of (o/n)first (o/n)second
}
\description{
\code{ComputeGC3syn} Computes the G+C content at the 1st and 2nd position of synonymous codons
}
\details{
Computes the GC content of the 1st and 2nd position of synonymous codons to estimate the background
GC content when compared with the GC content at the 3rd position. By definition, GC12s values are
the proportion of GC nucleotides at the fixed first and less variable second coding position of
synonymous codons. This can be used to evaluate the degree of base composition bias.
}
\examples{
ComputeGC12syn('ATGTGGTACTCCGACTACGGAGGATAA')
plot(mean(ComputeGC12syn(toupper(c2s(mylist(whatout=1)[[1]])))), ComputeGC3syn(toupper(c2s(mylist(whatout=1)[[1]]))))


}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}} \code{\link{ComputeGC3syn}}
}
\keyword{CodonBias}

