% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countcodonfreq.R
\name{countcodonfreq}
\alias{countcodonfreq}
\title{Count Codon Frequency}
\usage{
countcodonfreq(no = 1, wdir = "/windows/R/")
}
\arguments{
\item{no}{Number of file in working directory to be processed}

\item{wdir}{Adress pointing to directory with transcript and protein fasta files}
}
\value{
Codon Frequency Table
}
\description{
\code{countcodonfreq} Counts Codon Frequency for a Custom Genome (eg \emph{Eragostis tef})
}
\details{
This function should count the codon frequency in a genome based on a set of verified and matched transcript and protein fasta files.
This is a roxygen2 created helping file.
}
\examples{
countcodonfreq()

}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}}
}
\keyword{CodonBias}

