% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entries.R
\name{Entries}
\alias{Entries}
\title{Entries}
\usage{
Entries()
}
\value{
VOID
}
\description{
\code{Entries} DNA entries in the working data base
}
\details{
Should work the same as Entries() in Darwin on DB
}
\section{Original code in Darwin}{

\subsection{Compute CAI, the Codon Adaptation Index (Sharp and Li 1987)}{\preformatted{
end: } }
}
\examples{
Entries()

}
\author{
Roth, A.; Friberg, M.; Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}} \code{\link{CodonProbabilities}}
}
\keyword{CodonBias}

