% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindHighlyExpressedGenes.R
\name{FindHighlyExpressedGenes}
\alias{FindHighlyExpressedGenes}
\title{Find Highly Expressed Genes}
\usage{
FindHighlyExpressedGenes(n = 100, tag = "PROTEXPR")
}
\arguments{
\item{n}{Integer}

\item{tag}{String either 'PROTEXPR' or 'MRNAEXPR'}
}
\value{
VOID
}
\description{
\code{FindHighlyExpressedGenes} Finds Highly Expressed Genes in Lost
}
\details{
Should work the same as FindHighlyExpressedGenes() in Darwin on DB
}
\section{Original code in Darwin}{

\subsection{Compute CAI, the Codon Adaptation Index (Sharp and Li 1987)}{\preformatted{
FindHighlyExpressedGenes := proc( ; n=100:integer, tag='PROTEXPR':string)
 # tags: 'PROTEXPR' 'MRNAEXPR'
 expr := CreateArray(1..DB[TotEntries]);
 for i to DB[TotEntries] do
   ex := sscanf(SearchTag(tag, Entry(i)), '%f');
   if ex <> [] then
     expr[i] := op(ex) fi;
 od;
 sorted := sort(expr);
 limit := sorted[length(sorted)-n+1];
 genes := [];
 for i to DB[TotEntries] do
   if expr[i] >= limit then
     genes := append(genes, i) fi od;
   genes
end: } }
}
\examples{
FindHighlyExpressedGenes()

}
\author{
Roth, A.; Friberg, M.; Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}} \code{\link{CodonProbabilities}}
}
\keyword{CodonBias}

