% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GtRNAdb2species}
\alias{GtRNAdb2species}
\title{Viruses, Eukaryota, Archaea, Bacteria Species Names on GtRNAdb2}
\format{A vector of stings (partial file names)}
\source{
\url{http://gtrnadb.ucsc.edu/GtRNAdb2/}
}
\usage{
GtRNAdb2species
}
\description{
A Vector with Species Names from GtRNAdb, edited for html conformatibility
}
\keyword{datasets}

