% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverage.R
\name{movingAverage}
\alias{movingAverage}
\title{Calculating a Moving Average}
\usage{
movingAverage(x, n = 1, centered = FALSE)
}
\arguments{
\item{x}{the vector}

\item{n}{the number of samples}

\item{centered}{if FALSE, then average current sample and previous (n-1) samples
if TRUE, then average symmetrically in past and future. (If n is even, use one more sample from future.)}
}
\value{
sum divided by count
}
\description{
A way to calculate a moving average.
}
\details{
Suppose your data is a noisy sine wave with some missing values. A different way to handle missing data is to simply ignore it, and not include it in the average. The function defined here will do that.
}
\examples{
# Make same plots from before, with thicker lines
plot(x, y, type="l", col=grey(.5))
grid()
y_lag <- filter(y, rep(1/20, 20), sides=1)
lines(x, y_lag, col="red", lwd=4)         # Lagged average in red
y_sym <- filter(y, rep(1/21,21), sides=2)
lines(x, y_sym, col="blue", lwd=4)        # Symmetrical average in blue
# Calculate lagged moving average with new method and overplot with green
y_lag_na.rm <- movingAverage(y, 20)
lines(x, y_lag_na.rm, col="green", lwd=2)
y_sym_na.rm <- movingAverage(y, 21, TRUE)
lines(x, y_sym_na.rm, col="green", lwd=2)

}
\author{
Cookbook, R. \email{winston@stdout.org}
}
\seealso{
\url{http://www.cookbook-r.com/Manipulating_data/Calculating_a_moving_average/}
}

