% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOCU.R
\name{plotOCU}
\alias{plotOCU}
\title{Plot of Optimality Score for Codons in a Single Protein}
\usage{
plotOCU(entry = 3, isfreq = FALSE)
}
\arguments{
\item{entry}{Index number in SeqFastadna list myaa}

\item{isfreq}{frequencies are used instead of boolean of optimaity}
}
\value{
void
}
\description{
\code{plotOCU} Plots the Optimality Score for Codon in a Protein based on tRNA-Gene Abundancy in a Custom Genome (eg \emph{Eragostis tef})
}
\details{
Here a plot for the optimality of the used codons of a protein is plotted for its position in the protein sequence.
Two mooving averages are overlayed for better visuality.
In development stage!
}
\examples{
for (n in 1:length(mylist()[[2]]) try(plotOCU(n, TRUE))

}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{movingAverage}} \code{\link{statanacoseq}}
}
\keyword{CodonBias}

