% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfasta.R
\name{readfasta}
\alias{readfasta}
\title{Processed tRNA Containing Fasta File Headers to Count Anticodons}
\usage{
readfasta(fileno = 1, chunksize = 1, verbose = TRUE)
}
\arguments{
\item{verbose}{Wheter to show the aminoacid, anticodon and tRNA Model Score from fasta entry header}

\item{fam}{Taxonomy kingdom defining the folder on gtrnadb homepage}

\item{spec}{Species identifier as pointer to .fa filename}
}
\value{
tRNA gene count Table with statistics
}
\description{
\code{readfasta} processes fasta sequence headers from tRNA gene sequences of complete genomes and compares anticodon usage in the gene set.
}
\details{
Import helper function to manage automatical downloading of fasta files.
}
\examples{
readfasta(131)

}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}} \url{http://gtrnadb.ucsc.edu/GtRNAdb2/} \code{\link{readstats}}
}
\keyword{CodonBias}

