% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstats.R
\name{readstats}
\alias{readstats}
\title{Read tRNA Gene Count Statistics}
\usage{
readstats(fileno = 156, chunksize = 1)
}
\arguments{
\item{fileno}{Index number of html-file in the working directory}

\item{chunksize}{The maximum number of lines of data to be read}
}
\value{
tRNA gene count Table
}
\description{
\code{readstats} read-in anticodon count from tRNAscan-2.0 or similar summary files from GtRNAdb2 homepage
}
\details{
Puts gene count into a table according to anticodon sequence on tRNA.
This functions works but on the summary files there are some not
nicely behaving statistics where the new filtering methods have not been applied as for the fasta files.
}
\examples{
setwd("E:/R/viruses")
filecount <- length(dir(pattern=".stats.html$"))
vir_optcodon <- as.list(rep(data.frame(x=0), times=filecount))
for (n in 1:filecount) vir_optcodon[[n]] <- readstats(n)
names(vir_optcodon) <- sapply(dir(pattern="stats.html$"), function(x) strsplit(x, "-")[[1]][1])

}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}} \url{http://gtrnadb.ucsc.edu/GtRNAdb2/}
}
\keyword{CodonBias}

