% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtRNAout.R
\name{readtRNAout}
\alias{readtRNAout}
\title{Read tRNA Gene Count Statistics}
\usage{
readtRNAout(file = system.file("extdata", "seq30469.out", package =
  "statanacoseq"))
}
\arguments{
\item{file}{seq.out file from tRNAscan-SE 2.0}
}
\value{
tRNA gene count Table
}
\description{
\code{readtRNAout} read-in anticodon count from tRNAscan-2.0 output file from online web tool
}
\details{
Puts gene count into a table according to anticodon sequence on tRNA.
This functions works on a tRNA-scan output reanalyzed by tRNAscan-SE 2.0 online tool
to behave similarly to the data available on GtRNAdb2.
}
\examples{
readtRNAout()

}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}} \url{http://trna.ucsc.edu/tRNAscan-SE/}
}
\keyword{CodonBias}

