% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeAdaptiveness.R
\name{RelativeAdaptiveness}
\alias{RelativeAdaptiveness}
\title{Compute Relative Adaptiveness}
\usage{
RelativeAdaptiveness(entries = list(1))
}
\arguments{
\item{entires}{list of positive integers}
}
\value{
Named (codons) numerical vector with relative adaptiveness for the 64 codons
}
\description{
\code{RelativeAdaptiveness} calculates RA for a List of Entry Numbers
}
\details{
Should compute the same RA as in Darwin
}
\examples{
RelativeAdaptiveness()

}
\author{
Roth, A.; Friberg, M.; Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{ComputeCarboneRA}} \code{\link{statanacoseq}} \code{\link{SetupRA}}
}
\keyword{CodonBias}

