% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reversecomplement.R
\name{reversecomplement}
\alias{reversecomplement}
\title{Reverse Complement Anticodon}
\usage{
reversecomplement(codon)
}
\arguments{
\item{codon}{3 character string for codon}
}
\value{
upper case 3 character string for anticodon
}
\description{
\code{reversecomplement} calculates anticodon from reverse complement of its codon sequence
}
\details{
Reverse complement codon to isotype / anticodon sequence
}
\examples{
reversecomplement("CAT")

}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}} \url{http://trna.ucsc.edu/tRNAscan-SE/}
}
\keyword{CodonBias}

