% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSCU.R
\name{RSCU}
\alias{RSCU}
\title{Relative Synonymous Codon Usage}
\usage{
RSCU(cds)
}
\arguments{
\item{cds}{Coding sequence in reading frame}
}
\value{
Named (codons) numerical vector with relative synonymous codon usage for the 64 codons
}
\description{
\code{RSCU} Returns Relative Synonymous Codon Usage Similar tp SeqinR
}
\details{
Should compute the same RSCU as seqinr.
}
\section{Original code in Darwin}{

\subsection{Relative synonymous codon usage.  AR (2007)}{\preformatted{
RSCU := proc(;d:string)
  if nargs>0 then
    cc:=CodonCount(d);
  else
    cc := CodonCount();
  fi;
  rscu := CreateArray(1..64);
  for i to 64 do
    s:=0;
    syn:=IntToCInt(CIntToInt(i));
    l:=length(syn);
    for j in syn do
      s:=s+cc[j];
    od;
    if s=0 then next fi;
    rscu[i]:=cc[i]/(s/l);
  od;
  rscu;
end: } }
}
\examples{
RSCU('ATGTGGTACTCCGACTACGGAGGATAA')
RSCU(c2s(mylist(whatout=1)[[1]]))

}
\author{
Roth, A.; Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}} \code{\link{readstats}}
}
\keyword{CodonBias}

