% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaledChiSquare.R
\name{ScaledChiSquare}
\alias{ScaledChiSquare}
\title{Scaled Chi-Square}
\usage{
ScaledChiSquare(cds)
}
\arguments{
\item{cds}{Coding Sequence}
}
\value{
VOID
}
\description{
\code{ScaledChiSquare} Computes the Scaled Chi-Square Index for a Gene
}
\details{
Should be implemented because missing in Darwin
}
\section{Original code in Darwin}{

\subsection{Scaled Chi-Square.}{\preformatted{
ScaledChiSquare := proc(d:string)
# to be implemented
end: } }
}
\examples{
ScaledChiSquare('ATGTGGTACTCCGACTACGGAGGATAA')
ScaledChiSquare(c2s(mylist(whatout=1)[[1]]))

}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}}
}
\keyword{CodonBias}

