% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchfafile.R
\name{searchfafile}
\alias{searchfafile}
\title{Search and Return a Fasta File Name from Partial Name}
\usage{
searchfafile(fam = "eukaryota", spec = "Hsapi19")
}
\arguments{
\item{fam}{Taxonomy kingdom defining the folder on gtrnadb homepage}

\item{spec}{Species identifier as pointer to .fa filename}
}
\value{
string with .fa file name
}
\description{
\code{searchfafile} searches fasta file name on GtRNAdb2 homepage
}
\details{
Import helper function to manage automatical downloading of fasta files.
}
\examples{
for (i in virus) try(download.file(paste("http://gtrnadb.ucsc.edu/GtRNAdb2/genomes/viruses/", i, "/", searchfafile(fam='viruses', spec=i), sep=""), paste("viruses/", i, "-tRNAs.fa", sep="") ))

}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}} \url{http://gtrnadb.ucsc.edu/GtRNAdb2/}
}
\keyword{CodonBias}

