% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupRA.R
\name{SetupRA}
\alias{SetupRA}
\title{Relative Synonymous Codon Usage}
\usage{
SetupRA(s = "yeast")
}
\arguments{
\item{s}{Species and subsettings, maybe one of
c('yeast', 'yeast2perc', 'yeast1perc', 'yeast05perc', 'yeasttop24protexpr', 'yeasttop24mrnaexpr')
or 'carbone' for computing carbone RA with standard values.}
}
\value{
Numerical vector with relative adaptiveness for the 64 codons
}
\description{
\code{SetupRA } creates a table of relative adaptiveness for yeast or based on carbon RA
}
\details{
Generation of manually set tables.
}
\examples{
SetupRA('yeast2perc')

}
\author{
Roth, A.; Friberg, M.; Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{statanacoseq}} \code{\link{ComputeCarboneRA}}
}
\keyword{CodonBias}

