% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SilentSiteComposition.R
\name{SilentSiteComposition}
\alias{SilentSiteComposition}
\title{Base Composition at Silent Sites}
\usage{
SilentSiteComposition(cds)
}
\arguments{
\item{cds}{Coding Sequence for Silent Site Composition calculation}
}
\value{
VOID
}
\description{
\code{SilentSiteComposition} Computes the Base Composition at Silent Sites
}
\details{
Should be implemented because missing in Darwin
}
\section{Original code in Darwin}{

\subsection{Base composition at silent sites.}{\preformatted{
SilentSiteComposition := proc(d:string)
# to be implemented
end: } }
}
\examples{
SilentSiteComposition('ATGTGGTACTCCGACTACGGAGGATAA')
SilentSiteComposition(c2s(mylist(whatout=1)[[1]]))

}
\author{
Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}}
}
\keyword{CodonBias}

