% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tef}
\alias{Tef}
\title{Optimal Codon Table for Eragostis tef Decided on tRNAscan-SE count of tRNA genes in sequenced genome}
\format{A data frame with 64 rows (anticodons) and 6 descriptive variables:
\describe{
  \item{tRNAs}{Aminoacid coded by tRNA}
  \item{codons}{anticodon on tRNA gene}
  \item{reps}{Cove-confirmed tRNAs}
  \item{optfact}{# isotype (isoacceptor typ) / max(#) isotype for aminoacid}
  \item{frac}{# isotype / sum(#) isotype for aminoacid}
  \item{isopt}{boolean decision if it is (one of) the optimal codon - optfract == 1}
}}
\source{
\url{http://gtrnadb.ucsc.edu/GtRNAdb2/}
}
\usage{
Tef
}
\description{
A table with the number of tRNA genes in \emph{Eragostis tef} for optimal codon evaluation.
}
\keyword{datasets}

