% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{veab}
\alias{veab}
\title{Viruses, Eukaryota, Archaea, Bacteria Species File Names on GtRNAdb2}
\format{A vector of stings (file names)}
\source{
\url{http://gtrnadb.ucsc.edu/GtRNAdb2/}
}
\usage{
veab
}
\description{
A Vector with .html File Names Containing the tRNAscan-SE v.2.0 Analysis of Complete Genomes from GtRNAdb
}
\keyword{datasets}

