% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{veabfa}
\alias{veabfa}
\title{Viruses, Eukaryota, Archaea, Bacteria Species Fasta File Names on GtRNAdb2}
\format{A vector of stings (file names)}
\source{
\url{http://gtrnadb.ucsc.edu/GtRNAdb2/}
}
\usage{
veabfa
}
\description{
A Vector with .fa File Names Containing the tRNA Gene List of Complete Genomes from GtRNAdb
}
\keyword{datasets}

