% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeGC3syn.R
\name{ComputeGC3syn}
\alias{ComputeGC3syn}
\title{GC content 3rd position of synonymous codons}
\usage{
ComputeGC3syn(tD)
}
\arguments{
\item{td}{Dictionary of String and Entry # nope probably either string or Entry type}

\item{pos}{List of 3 positive integers for the nucleotide positions}
}
\value{
o/n
}
\description{
\code{ComputeGC3syn} Computes the G+C content 3rd position of synonymous codons
}
\details{
Should compute the same CP as in Darwin
}
\section{Original code in Darwin}{

\subsection{G+C content 3rd position of synonymous codons.  AR (April 2007)}{\preformatted{
ComputeGC3syn:= proc(td:string)
 if member(td[-3..-1], AToCodon('$')) then d:=td[1..-4] else d:=td fi; # remove stop codon
 o := CreateArray(1..4);
 n:=0;
 for i to length(d) by 3 do
   c:=d[i..i+2];
   if length(IntToCInt(CodonToInt(c)))>1 then
     n:=n+1;
     oi:=BToInt(c[3]);
     o[oi] := o[oi]+1
   fi;
 od;
 o:=o/n;
 return(o[2]+o[3]);
end: } }
}
\examples{
ComputeGC3syn('ATGTGGTACTCCGACTACGGAGGATAA')
ComputeGC3syn(c2s(mylist(whatout=1)[[1]]))

}
\author{
Roth, A.; Friberg, M.; Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}} \code{\link{readstats}}
}
\keyword{CodonBias}

