% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeNEC.R
\name{ComputeNEC}
\alias{ComputeNEC}
\title{Effective number of codons}
\usage{
ComputeNEC(cds)
}
\arguments{
\item{cds}{Coding sequence in reading frame}
}
\value{
Numerical value for effective number of codons Nc (by the original definition)
}
\description{
\code{ComputeNEC} is a method for measuring synonymous codon usage bias in a gene by estimation of the “effective number of codons” Nc
}
\details{
Based on early codon usage mesures proposed by Wright 1990 and Fuglsang 2004. Original implementation from Alexander Roth 2005-2007,
second of several indices for codon usage. Nc is the total number of different codons used in a sequence. The values for Nc range from 20 to 61.
Where in the first case only one codon is used per amino acid, and in the latter all posible ones are used.
Highly expressed genes should use fewer codons due to selection away from equal use. The here computed overall number of effective codons for a gene
is a sum of average homozygosities for different redundency classes.
}
\examples{
ComputeNEC('ATGTGGTACTCCGACTACGGAGGATAA')
ComputeNEC2(c2s(mylist(whatout=1)[[1]]))

}
\author{
Roth, A.; Siegrist, F. and Cannarozzi, G. M. \email{gina@cannarozzi.com}
}
\seealso{
\code{\link{seqinr}} \code{\link{statanacoseq}} \url{http://gtrnadb.ucsc.edu/GtRNAdb2/} \code{\link{readstats}}
}
\keyword{CodonBias}

